+++ packages/core/src/presets/debridioScraper.ts
@@ -18,6 +18,7 @@ export class DebridioPreset extends Preset {
       constants.PREMIUMIZE_SERVICE,
       constants.TORBOX_SERVICE,
       constants.EASYDEBRID_SERVICE,
+      constants.DEBRIDER_SERVICE,
     ];
     const supportedResources = [constants.STREAM_RESOURCE];
 
@@ -35,6 +36,7 @@ export class DebridioPreset extends Preset {
           'Optionally override the services that are used. If not specified, then the services that are enabled and supported will be used.',
         type: 'multi-select',
         required: false,
+        showInNoobMode: false,
         options: supportedServices.map((service) => ({
           value: service,
           label: constants.SERVICE_DETAILS[service].name,