+++ packages/core/src/utils/config.ts
@@ -245,7 +245,9 @@ export function getEnvironmentServiceDetails(): typeof constants.SERVICE_DETAILS
             name: cred.name,
             description: cred.description,
             type: cred.type,
-            required: cred.required,
+            // remove required attribute from field to allow users to remove credentials.
+            // server will still validate.
+            required: false,
             default: getServiceCredentialDefault(service.id, cred.id)
               ? encryptString(getServiceCredentialDefault(service.id, cred.id)!)
                   .data
@@ -254,6 +256,9 @@ export function getEnvironmentServiceDetails(): typeof constants.SERVICE_DETAILS
               ? encryptString(getServiceCredentialForced(service.id, cred.id)!)
                   .data
               : null,
+            constraints: {
+              min: 1,
+            },
           })),
         },
       ])
@@ -412,10 +417,9 @@ export async function validateConfig(
 
   await validateRegexes(config);
 
-  await new AIOStreams(
-    ensureDecrypted(config),
-    skipErrorsFromAddonsOrProxies
-  ).initialise();
+  await new AIOStreams(ensureDecrypted(config), {
+    skipFailedAddons: skipErrorsFromAddonsOrProxies,
+  }).initialise();
 
   return config;
 }
@@ -491,7 +495,7 @@ async function validateRegexes(config: UserData) {
   const includedRegexes = config.includedRegexPatterns;
   const requiredRegexes = config.requiredRegexPatterns;
   const preferredRegexes = config.preferredRegexPatterns;
-  const regexAllowed = FeatureControl.isRegexAllowed(config);
+  const regexAllowed = await FeatureControl.isRegexAllowed(config);
 
   const regexes = [
     ...(excludedRegexes ?? []),
@@ -649,6 +653,9 @@ function validateOption(
   value: any,
   decryptValues: boolean = false
 ): any {
+  if (typeof value === 'string' && value === 'undefined') {
+    value = undefined;
+  }
   const forcedValue =
     option.forced !== undefined && option.forced !== null
       ? option.forced